requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
langdetect==1.0.9
libretranslatepy==2.1.1
lxml==4.9.2
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
openai==0.27.7
packaging==23.1
PyAudio==0.2.13
pydub==0.25.1
pypiwin32==223
pyttsx3==2.90
pywin32==306
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
semantic-version==2.10.0
setuptools-rust==1.6.0
six==1.16.0
SpeechRecognition==3.10.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
translate==3.6.1
typing_extensions==4.5.0
urllib3==2.0.2
Wave==0.0.2
whisper-openai==1.0.0
yarl==1.9.2
